+++ Sources/View/Channel/MessageCell/NotificationChannel/SBUChatNotificationCell.swift
@@ -14,6 +14,27 @@ protocol SBUChatNotificationCellDelegate: SBUNotificationCellDelegate {}
 class SBUChatNotificationCell: SBUNotificationCell {
     override func setupViews() {
         self.type = .chat
+        self.topMargin = 16
+        self.bottomMargin = 0
         super.setupViews()
+    }
+    
+    override func setupLayouts() {
+        super.setupLayouts()
+
+        if let topMarginConstraint = self.topMarginConstraint {
+            self.contentView.removeConstraint(topMarginConstraint)
+        }
+        if let bottomMarginConstraint = self.bottomMarginConstraint {
+            self.contentView.removeConstraint(bottomMarginConstraint)
+        }
+        self.topMarginConstraint?.isActive = false
+        self.bottomMarginConstraint?.isActive = false
+        self.topMarginConstraint = self.stackView.topAnchor.constraint(equalTo: self.contentView.topAnchor, constant: self.topMargin)
+        self.bottomMarginConstraint = self.stackView.bottomAnchor.constraint(equalTo: self.contentView.bottomAnchor, constant: self.bottomMargin)
+        self.topMarginConstraint?.isActive = true
+        self.bottomMarginConstraint?.isActive = true
+        self.contentView.updateConstraints()
+        self.contentView.layoutIfNeeded()
     }
 }